+++ source/Example.Orchestrations.Tests/Fixtures/ExampleOrchestrationsAppManager.cs
@@ -1,0 +1,240 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System.Diagnostics;
+using System.Diagnostics.CodeAnalysis;
+using Energinet.DataHub.Core.DurableFunctionApp.TestCommon.DurableTask;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.Azurite;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.Configuration;
+using Energinet.DataHub.Core.FunctionApp.TestCommon.FunctionAppHost;
+using Energinet.DataHub.Core.TestCommon.Diagnostics;
+using Energinet.DataHub.ProcessManagement.Core.Infrastructure.Extensions.Options;
+using Energinet.DataHub.ProcessManager.Core.Tests.Fixtures;
+using Microsoft.Azure.WebJobs.Extensions.DurableTask;
+using Xunit.Abstractions;
+
+namespace Energinet.DataHub.Example.Orchestrations.Tests.Fixtures;
+
+/// <summary>
+/// Support testing Example Orchestrations app and specifying configuration using inheritance.
+/// This allows us to use multiple fixtures and coordinate their configuration.
+/// </summary>
+public class ExampleOrchestrationsAppManager : IAsyncDisposable
+{
+    /// <summary>
+    /// Durable Functions Task Hub Name
+    /// See naming constraints: https://learn.microsoft.com/en-us/azure/azure-functions/durable/durable-functions-task-hubs?tabs=csharp#task-hub-names
+    /// </summary>
+    private readonly string _taskHubName;
+
+    private readonly int _appPort;
+    private readonly bool _manageDatabase;
+    private readonly bool _manageAzurite;
+
+    public ExampleOrchestrationsAppManager()
+        : this(
+            new ProcessManagerDatabaseManager("ExampleOrchestrationsTest"),
+            new IntegrationTestConfiguration(),
+            new AzuriteManager(useOAuth: true),
+            taskHubName: "ExampleOrchestrationsTest01",
+            appPort: 8012,
+            manageDatabase: true,
+            manageAzurite: true)
+    {
+    }
+
+    public ExampleOrchestrationsAppManager(
+        ProcessManagerDatabaseManager databaseManager,
+        IntegrationTestConfiguration configuration,
+        AzuriteManager azuriteManager,
+        string taskHubName,
+        int appPort,
+        bool manageDatabase,
+        bool manageAzurite)
+    {
+        _taskHubName = string.IsNullOrWhiteSpace(taskHubName)
+            ? throw new ArgumentException("Cannot be null or whitespace.", nameof(taskHubName))
+            : taskHubName;
+        _appPort = appPort;
+        _manageDatabase = manageDatabase;
+        _manageAzurite = manageAzurite;
+
+        DatabaseManager = databaseManager;
+        TestLogger = new TestDiagnosticsLogger();
+
+        IntegrationTestConfiguration = configuration;
+        AzuriteManager = azuriteManager;
+
+        DurableTaskManager = new DurableTaskManager(
+            nameof(ProcessManagerTaskHubOptions.ProcessManagerStorageConnectionString),
+            AzuriteManager.FullConnectionString);
+    }
+
+    public ProcessManagerDatabaseManager DatabaseManager { get; }
+
+    public ITestDiagnosticsLogger TestLogger { get; }
+
+    [NotNull]
+    public FunctionAppHostManager? AppHostManager { get; private set; }
+
+    [NotNull]
+    public IDurableClient? DurableClient { get; private set; }
+
+    private IntegrationTestConfiguration IntegrationTestConfiguration { get; }
+
+    private AzuriteManager AzuriteManager { get; }
+
+    private DurableTaskManager DurableTaskManager { get; }
+
+    /// <summary>
+    /// Start the example orchestration app
+    /// </summary>
+    public async Task StartAsync()
+    {
+        if (_manageAzurite)
+        {
+            // Clean up old Azurite storage
+            CleanupAzuriteStorage();
+
+            // Storage emulator
+            AzuriteManager.StartAzurite();
+        }
+
+        if (_manageDatabase)
+            await DatabaseManager.CreateDatabaseAsync();
+
+        // Prepare host settings
+        var appHostSettings = CreateAppHostSettings("Example.Orchestrations");
+
+        // Create and start host
+        AppHostManager = new FunctionAppHostManager(appHostSettings, TestLogger);
+
+        StartHost(AppHostManager);
+
+        DurableClient = DurableTaskManager.CreateClient(taskHubName: _taskHubName);
+    }
+
+    public async ValueTask DisposeAsync()
+    {
+        AppHostManager.Dispose();
+        await DurableTaskManager.DisposeAsync();
+
+        if (_manageAzurite)
+            AzuriteManager.Dispose();
+
+        if (_manageDatabase)
+            await DatabaseManager.DeleteDatabaseAsync();
+    }
+
+    /// <summary>
+    /// Use this method to attach <paramref name="testOutputHelper"/> to the host logging pipeline.
+    /// While attached, any entries written to host log pipeline will also be logged to xUnit test output.
+    /// It is important that it is only attached while a test is active. Hence, it should be attached in
+    /// the test class constructor; and detached in the test class Dispose method (using 'null').
+    /// </summary>
+    /// <param name="testOutputHelper">If a xUnit test is active, this should be the instance of xUnit's <see cref="ITestOutputHelper"/>;
+    /// otherwise it should be 'null'.</param>
+    public void SetTestOutputHelper(ITestOutputHelper? testOutputHelper)
+    {
+        TestLogger.TestOutputHelper = testOutputHelper;
+    }
+
+    /// <summary>
+    /// Cleanup Azurite storage to avoid situations where Durable Functions
+    /// would otherwise continue working on old orchestrations that e.g. failed in
+    /// previous runs.
+    /// </summary>
+    public void CleanupAzuriteStorage()
+    {
+        AzuriteManager.CleanupAzuriteStorage();
+    }
+
+    private static void StartHost(FunctionAppHostManager hostManager)
+    {
+        IEnumerable<string> hostStartupLog;
+
+        try
+        {
+            hostManager.StartHost();
+        }
+        catch (Exception)
+        {
+            // Function App Host failed during startup.
+            // Exception has already been logged by host manager.
+            hostStartupLog = hostManager.GetHostLogSnapshot();
+
+            if (Debugger.IsAttached)
+                Debugger.Break();
+
+            // Rethrow
+            throw;
+        }
+
+        // Function App Host started.
+        hostStartupLog = hostManager.GetHostLogSnapshot();
+    }
+
+    private static string GetBuildConfiguration()
+    {
+#if DEBUG
+        return "Debug";
+#else
+        return "Release";
+#endif
+    }
+
+    private FunctionAppHostSettings CreateAppHostSettings(string csprojName)
+    {
+        var buildConfiguration = GetBuildConfiguration();
+
+        var appHostSettings = new FunctionAppHostConfigurationBuilder()
+            .CreateFunctionAppHostSettings();
+
+        appHostSettings.FunctionApplicationPath = $"..\\..\\..\\..\\{csprojName}\\bin\\{buildConfiguration}\\net8.0";
+        appHostSettings.Port = _appPort;
+
+        // It seems the host + worker is not ready if we use the default startup log message, so we override it here
+        appHostSettings.HostStartedEvent = "Host lock lease acquired";
+
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            "FUNCTIONS_WORKER_RUNTIME",
+            "dotnet-isolated");
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            "AzureWebJobsStorage",
+            AzuriteManager.FullConnectionString);
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            "APPLICATIONINSIGHTS_CONNECTION_STRING",
+            IntegrationTestConfiguration.ApplicationInsightsConnectionString);
+        // Make Orchestrator poll for updates every second (default is every 30 seconds) by overriding maxQueuePollingInterval
+        // (ref: https://learn.microsoft.com/en-us/azure/azure-functions/durable/durable-functions-bindings?tabs=python-v2%2Cisolated-process%2C2x-durable-functions&pivots=programming-language-csharp#hostjson-settings)
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            "AzureFunctionsJobHost__extensions__durableTask__storageProvider__maxQueuePollingInterval",
+            "00:00:01");
+
+        // ProcessManager
+        // => Task Hub
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            nameof(ProcessManagerTaskHubOptions.ProcessManagerStorageConnectionString),
+            AzuriteManager.FullConnectionString);
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            nameof(ProcessManagerTaskHubOptions.ProcessManagerTaskHubName),
+            _taskHubName);
+        // => Database
+        appHostSettings.ProcessEnvironmentVariables.Add(
+            $"{ProcessManagerOptions.SectionName}__{nameof(ProcessManagerOptions.SqlDatabaseConnectionString)}",
+            DatabaseManager.ConnectionString);
+
+        return appHostSettings;
+    }
+}