+++ source/ProcessManager.Orchestrations/Processes/BRS_023_027/V1/Activities/EnqueueMessagesStepTerminateActivity_Brs_023_027_V1.cs
@@ -28,20 +28,22 @@ internal class EnqueueMessagesStepTerminateActivity_Brs_023_027_V1(
 {
     [Function(nameof(EnqueueMessagesStepTerminateActivity_Brs_023_027_V1))]
     public async Task Run(
-        [ActivityTrigger] Guid orchestrationInstanceId)
+        [ActivityTrigger] ActivityInput input)
     {
         var orchestrationInstance = await ProgressRepository
-            .GetAsync(new OrchestrationInstanceId(orchestrationInstanceId))
+            .GetAsync(input.InstanceId)
             .ConfigureAwait(false);
 
-        var step = orchestrationInstance.Steps.Single(x => x.Sequence == Orchestration_Brs_023_027_V1.EnqueueMessagesStepSequence);
-        if (!step.IsSkipped())
-        {
-            step.Lifecycle.TransitionToTerminated(Clock, OrchestrationStepTerminationStates.Succeeded);
-            await ProgressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);
+        var step = orchestrationInstance.Steps
+            .Single(step => step.Sequence == Orchestration_Brs_023_027_V1.EnqueueMessagesStepSequence);
 
-            // TODO: For demo purposes; remove when done
-            await Task.Delay(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
-        }
+        step.Lifecycle.TransitionToTerminated(Clock, OrchestrationStepTerminationStates.Succeeded);
+        await ProgressRepository.UnitOfWork.CommitAsync().ConfigureAwait(false);
+
+        // TODO: For demo purposes; remove when done
+        await Task.Delay(TimeSpan.FromSeconds(1)).ConfigureAwait(false);
     }
+
+    public record ActivityInput(
+        OrchestrationInstanceId InstanceId);
 }