+++ docs/ProcessManager.Orchestrations.Abstractions/ReleaseNotes/ReleaseNotes.md
@@ -5,6 +5,10 @@
 - Do not allow empty string in values of the following options:
     - `ProcessManagerTopicOptions`
 
+## Version 0.2.5
+
+- Add actor id to start orchestration messages.
+
 ## Version 0.2.4
 
 - Update (implement) correct input for starting BRS-026 and BRS-028 orchestrations.